+++ stroom-core-server/src/main/java/stroom/node/server/DefaultProperties.java
@@ -973,12 +973,6 @@ public class DefaultProperties {
                 .description("A single regex that will be used to replace all matches in the class name with '--REPLACED--'. This is to prevent ids for anonymous inner classes and lambdas from being included in the class name. E.g '....DocRefResourceHttpClient$$Lambda$46/1402766141' becomes '....DocRefResourceHttpClient$$Lambda$--REPLACED--'. ")
                 .editable(true)
                 .build());
-        list.add(new GlobalProperty.Builder()
-                .name("stroom.node.status.heapHistogram.jMapExecutable")
-                .value("jmap")
-                .description("The jmap executable name if it is available on the PATH or a fully qualified form")
-                .build());
-
 
         // External Service properties, including DocRef.type name mappings
         // ========================================START===========================================